feat(runtime): implement Codex app-server startup with retry logic and enhanced logging
